Consolidate scores wire parsers and use camelCase segment fields.
Shared isRecord and readInferenceSolution module removes duplication; accelerated segment parsing matches backend camelCase contract.
