+++ packages/frontend/src/analytics/scores/acceleratedInferenceSegments.ts
@@ -1,9 +1,5 @@
 import type { ScoresInferenceSolution } from '../../api/bff'
-import { readMilitaryScoreArithmetic } from './inferenceConstraints'
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return value != null && typeof value === 'object' && !Array.isArray(value)
-}
+import { isRecord, readInferenceSolution } from './scoresWireParsers'
 
 export type AcceleratedInferenceSegment = {
   segmentId: string
@@ -16,53 +12,10 @@ export type AcceleratedInferenceSegment = {
   solutions: ScoresInferenceSolution[]
 }
 
-function readSolutionAction(entry: unknown): ScoresInferenceSolution['actions'][number] | null {
-  if (!isRecord(entry)) {
-    return null
-  }
-  if (
-    typeof entry.actionId !== 'string' ||
-    typeof entry.label !== 'string' ||
-    typeof entry.count !== 'number'
-  ) {
-    return null
-  }
-  return {
-    actionId: entry.actionId,
-    label: entry.label,
-    count: entry.count,
-  }
-}
-
-function readInferenceSolution(entry: unknown): ScoresInferenceSolution | null {
-  if (!isRecord(entry)) {
-    return null
-  }
-  if (typeof entry.objectiveValue !== 'number') {
-    return null
-  }
-  const actionsRaw = entry.actions
-  const actions: ScoresInferenceSolution['actions'] = []
-  if (Array.isArray(actionsRaw)) {
-    for (const action of actionsRaw) {
-      const parsed = readSolutionAction(action)
-      if (parsed != null) {
-        actions.push(parsed)
-      }
-    }
-  }
-  const arithmetic = readMilitaryScoreArithmetic(entry.militaryScoreArithmetic)
-  return {
-    objectiveValue: entry.objectiveValue,
-    actions,
-    ...(arithmetic != null ? { militaryScoreArithmetic: arithmetic } : {}),
-  }
-}
-
 export function readAcceleratedInferenceSegments(
   diagnostics: Record<string, unknown>
 ): AcceleratedInferenceSegment[] | null {
-  const raw = diagnostics.accelerated_segments ?? diagnostics.acceleratedSegments
+  const raw = diagnostics.accelerated_segments
   if (!Array.isArray(raw) || raw.length === 0) {
     return null
   }
@@ -72,19 +25,7 @@ export function readAcceleratedInferenceSegments(
     if (!isRecord(entry)) {
       continue
     }
-    const segmentId =
-      typeof entry.segmentId === 'string'
-        ? entry.segmentId
-        : typeof entry.segment_id === 'string'
-          ? entry.segment_id
-          : null
-    const hostTurn =
-      typeof entry.hostTurn === 'number'
-        ? entry.hostTurn
-        : typeof entry.host_turn === 'number'
-          ? entry.host_turn
-          : null
-    if (segmentId == null || hostTurn == null) {
+    if (typeof entry.segmentId !== 'string' || typeof entry.hostTurn !== 'number') {
       continue
     }
     const solutionsRaw = entry.solutions
@@ -98,33 +39,14 @@ export function readAcceleratedInferenceSegments(
       }
     }
     segments.push({
-      segmentId,
-      hostTurn,
+      segmentId: entry.segmentId,
+      hostTurn: entry.hostTurn,
       status: typeof entry.status === 'string' ? entry.status : 'unknown',
       solutionCount:
-        typeof entry.solutionCount === 'number'
-          ? entry.solutionCount
-          : typeof entry.solution_count === 'number'
-            ? entry.solution_count
-            : solutions.length,
-      militaryDelta2x:
-        typeof entry.militaryDelta2x === 'number'
-          ? entry.militaryDelta2x
-          : typeof entry.military_delta_2x === 'number'
-            ? entry.military_delta_2x
-            : 0,
-      warshipDelta:
-        typeof entry.warshipDelta === 'number'
-          ? entry.warshipDelta
-          : typeof entry.warship_delta === 'number'
-            ? entry.warship_delta
-            : 0,
-      freighterDelta:
-        typeof entry.freighterDelta === 'number'
-          ? entry.freighterDelta
-          : typeof entry.freighter_delta === 'number'
-            ? entry.freighter_delta
-            : 0,
+        typeof entry.solutionCount === 'number' ? entry.solutionCount : solutions.length,
+      militaryDelta2x: typeof entry.militaryDelta2x === 'number' ? entry.militaryDelta2x : 0,
+      warshipDelta: typeof entry.warshipDelta === 'number' ? entry.warshipDelta : 0,
+      freighterDelta: typeof entry.freighterDelta === 'number' ? entry.freighterDelta : 0,
       solutions,
     })
   }
